Factor out large generic additions in System.Linq
New generic types in System.Linq are causing significant size-on-disk increases in AOT builds.  This change addresses some of them:

- The many implementations of IPartition and IIListProvider are causing significant binary size increases for AOT.  I've factored out the majority of those large changes to only be used in a netcoreapp build, not in a uap build.  There are still code paths that test whether an object is IPartition or IIListProvider, but there are no types now in the UAP build that implement them.
- As part of this, I replaced several places where `EmptyPartition<TResult>.Instance` was being used directly to implement an `IEnumerable<T>`, changed them to use `Enumerable.Empty<T>`, and then changed `Enumerable.Empty<T>` so that on netcoreapp it returns `EmptyPartition<TResult>.Instance`, and on uap it returns `Array.Empty<T>`.  This also means now that code paths lighting-up on partitions will also light-up on `Enumerable.Empty<T>`.  It also means that I changed the empty partition's `IEnumerator.Reset` method to be a nop instead of throwing, to match the same behavior as array's `IEnumerator.Reset`.
- LargeArrayBuilder uses T[][], which introduces lots of new array types and corresponding interface instantiations, helper types, etc.  I've specialized LargeArrayBuilder in the UAP build to just be a thin wrapper around ArrayBuilder; we still end up with unnecessary instantiations of LargeArrayBuilder, but they're relatively small, and it saves everything to do with those T[][] instantiations.
