+++ src/System.Linq/src/System.Linq.csproj
@@ -10,6 +10,37 @@
       <TargetGroup>netcoreapp</TargetGroup>
     </ContractProject>
   </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp'">
+    <Compile Include="System\Linq\AppendPrepend.netcoreapp.cs" />
+    <Compile Include="System\Linq\Concat.netcoreapp.cs" />
+    <Compile Include="System\Linq\DefaultIfEmpty.netcoreapp.cs" />
+    <Compile Include="System\Linq\Distinct.netcoreapp.cs" />
+    <Compile Include="System\Linq\Enumerable.netcoreapp.cs" />
+    <Compile Include="System\Linq\Grouping.netcoreapp.cs" />
+    <Compile Include="System\Linq\Lookup.netcoreapp.cs" />
+    <Compile Include="System\Linq\OrderedEnumerable.netcoreapp.cs" />
+    <Compile Include="System\Linq\Partition.netcoreapp.cs" />
+    <Compile Include="System\Linq\Range.netcoreapp.cs" />
+    <Compile Include="System\Linq\Repeat.netcoreapp.cs" />
+    <Compile Include="System\Linq\Reverse.netcoreapp.cs" />
+    <Compile Include="System\Linq\Select.netcoreapp.cs" />
+    <Compile Include="System\Linq\SelectMany.netcoreapp.cs" />
+    <Compile Include="System\Linq\Skip.netcoreapp.cs" />
+    <Compile Include="System\Linq\Take.netcoreapp.cs" />
+    <Compile Include="System\Linq\Union.netcoreapp.cs" />
+    <Compile Include="System\Linq\Where.netcoreapp.cs" />
+    <Compile Include="$(CommonPath)\System\Collections\Generic\LargeArrayBuilder.netcoreapp.cs">
+      <Link>System\Collections\Generic\LargeArrayBuilder.netcoreapp.cs</Link>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'uap'">
+    <Compile Include="System\Linq\Enumerable.uap.cs" />
+    <Compile Include="System\Linq\Skip.uap.cs" />
+    <Compile Include="System\Linq\Take.uap.cs" />
+    <Compile Include="$(CommonPath)\System\Collections\Generic\LargeArrayBuilder.uap.cs">
+      <Link>System\Collections\Generic\LargeArrayBuilder.uap.cs</Link>
+    </Compile>
+  </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp' or '$(TargetGroup)' == 'uap'">
     <Compile Include="$(CommonPath)\System\Collections\Generic\ArrayBuilder.cs">
       <Link>System\Collections\Generic\ArrayBuilder.cs</Link>
@@ -47,6 +78,8 @@
     <Compile Include="System\Linq\GroupJoin.cs" />
     <Compile Include="System\Linq\Intersect.cs" />
     <Compile Include="System\Linq\Iterator.cs" />
+    <Compile Include="System\Linq\IIListProvider.cs" />
+    <Compile Include="System\Linq\IPartition.cs" />
     <Compile Include="System\Linq\Join.cs" />
     <Compile Include="System\Linq\Last.cs" />
     <Compile Include="System\Linq\Lookup.cs" />
@@ -54,7 +87,6 @@
     <Compile Include="System\Linq\Min.cs" />
     <Compile Include="System\Linq\OrderBy.cs" />
     <Compile Include="System\Linq\OrderedEnumerable.cs" />
-    <Compile Include="System\Linq\Partition.cs" />
     <Compile Include="System\Linq\Range.cs" />
     <Compile Include="System\Linq\Repeat.cs" />
     <Compile Include="System\Linq\Reverse.cs" />
