+++ src/System.Linq/src/System/Linq/AppendPrepend.cs
@@ -37,7 +37,7 @@ namespace System.Linq
         /// Represents the insertion of one or more items before or after an <see cref="IEnumerable{TSource}"/>.
         /// </summary>
         /// <typeparam name="TSource">The type of the source enumerable.</typeparam>
-        private abstract class AppendPrependIterator<TSource> : Iterator<TSource>, IIListProvider<TSource>
+        private abstract partial class AppendPrependIterator<TSource> : Iterator<TSource>
         {
             protected readonly IEnumerable<TSource> _source;
             protected IEnumerator<TSource> _enumerator;
@@ -80,19 +80,13 @@ namespace System.Linq
 
                 base.Dispose();
             }
-
-            public abstract TSource[] ToArray();
-
-            public abstract List<TSource> ToList();
-
-            public abstract int GetCount(bool onlyIfCheap);
         }
 
         /// <summary>
         /// Represents the insertion of an item before or after an <see cref="IEnumerable{TSource}"/>.
         /// </summary>
         /// <typeparam name="TSource">The type of the source enumerable.</typeparam>
-        private class AppendPrepend1Iterator<TSource> : AppendPrependIterator<TSource>
+        private partial class AppendPrepend1Iterator<TSource> : AppendPrependIterator<TSource>
         {
             private readonly TSource _item;
             private readonly bool _appending;
@@ -165,93 +159,13 @@ namespace System.Linq
                     return new AppendPrependN<TSource>(_source, new SingleLinkedNode<TSource>(_item).Add(item), null, prependCount: 2, appendCount: 0);
                 }
             }
-
-            private TSource[] LazyToArray()
-            {
-                Debug.Assert(GetCount(onlyIfCheap: true) == -1);
-
-                var builder = new LargeArrayBuilder<TSource>(initialize: true);
-                
-                if (!_appending)
-                {
-                    builder.SlowAdd(_item);
-                }
-
-                builder.AddRange(_source);
-
-                if (_appending)
-                {
-                    builder.SlowAdd(_item);
-                }
-
-                return builder.ToArray();
-            }
-
-            public override TSource[] ToArray()
-            {
-                int count = GetCount(onlyIfCheap: true);
-                if (count == -1)
-                {
-                    return LazyToArray();
-                }
-
-                TSource[] array = new TSource[count];
-                int index;
-                if (_appending)
-                {
-                    index = 0;
-                }
-                else
-                {
-                    array[0] = _item;
-                    index = 1;
-                }
-
-                EnumerableHelpers.Copy(_source, array, index, count - 1);
-
-                if (_appending)
-                {
-                    array[array.Length - 1] = _item;
-                }
-
-                return array;
-            }
-
-            public override List<TSource> ToList()
-            {
-                int count = GetCount(onlyIfCheap: true);
-                List<TSource> list = count == -1 ? new List<TSource>() : new List<TSource>(count);
-                if (!_appending)
-                {
-                    list.Add(_item);
-                }
-
-                list.AddRange(_source);
-                if (_appending)
-                {
-                    list.Add(_item);
-                }
-
-                return list;
-            }
-
-            public override int GetCount(bool onlyIfCheap)
-            {
-                if (_source is IIListProvider<TSource> listProv)
-                {
-                    int count = listProv.GetCount(onlyIfCheap);
-                    return count == -1 ? -1 : count + 1;
-                }
-
-                return !onlyIfCheap || _source is ICollection<TSource> ? _source.Count() + 1 : -1;
-            }
         }
 
         /// <summary>
         /// Represents the insertion of multiple items before or after an <see cref="IEnumerable{TSource}"/>.
         /// </summary>
         /// <typeparam name="TSource">The type of the source enumerable.</typeparam>
-        private class AppendPrependN<TSource> : AppendPrependIterator<TSource>
+        private partial class AppendPrependN<TSource> : AppendPrependIterator<TSource>
         {
             private readonly SingleLinkedNode<TSource> _prepended;
             private readonly SingleLinkedNode<TSource> _appended;
@@ -327,113 +241,6 @@ namespace System.Linq
             {
                 var prepended = _prepended != null ? _prepended.Add(item) : new SingleLinkedNode<TSource>(item);
                 return new AppendPrependN<TSource>(_source, prepended, _appended, _prependCount + 1, _appendCount);
-            }
-
-            private TSource[] LazyToArray()
-            {
-                Debug.Assert(GetCount(onlyIfCheap: true) == -1);
-
-                var builder = new SparseArrayBuilder<TSource>(initialize: true);
-
-                if (_prepended != null)
-                {
-                    builder.Reserve(_prependCount);
-                }
-
-                builder.AddRange(_source);
-
-                if (_appended != null)
-                {
-                    builder.Reserve(_appendCount);
-                }
-
-                TSource[] array = builder.ToArray();
-
-                int index = 0;
-                for (SingleLinkedNode<TSource> node = _prepended; node != null; node = node.Linked)
-                {
-                    array[index++] = node.Item;
-                }
-
-                index = array.Length - 1;
-                for (SingleLinkedNode<TSource> node = _appended; node != null; node = node.Linked)
-                {
-                    array[index--] = node.Item;
-                }
-
-                return array;
-            }
-
-            public override TSource[] ToArray()
-            {
-                int count = GetCount(onlyIfCheap: true);
-                if (count == -1)
-                {
-                    return LazyToArray();
-                }
-
-                TSource[] array = new TSource[count];
-                int index = 0;
-                for (SingleLinkedNode<TSource> node = _prepended; node != null; node = node.Linked)
-                {
-                    array[index] = node.Item;
-                    ++index;
-                }
-
-                if (_source is ICollection<TSource> sourceCollection)
-                {
-                    sourceCollection.CopyTo(array, index);
-                }
-                else
-                {
-                    foreach (TSource item in _source)
-                    {
-                        array[index] = item;
-                        ++index;
-                    }
-                }
-
-                index = array.Length;
-                for (SingleLinkedNode<TSource> node = _appended; node != null; node = node.Linked)
-                {
-                    --index;
-                    array[index] = node.Item;
-                }
-
-                return array;
-            }
-
-            public override List<TSource> ToList()
-            {
-                int count = GetCount(onlyIfCheap: true);
-                List<TSource> list = count == -1 ? new List<TSource>() : new List<TSource>(count);
-                for (SingleLinkedNode<TSource> node = _prepended; node != null; node = node.Linked)
-                {
-                    list.Add(node.Item);
-                }
-
-                list.AddRange(_source);
-                if (_appended != null)
-                {
-                    IEnumerator<TSource> e = _appended.GetEnumerator(_appendCount);
-                    while (e.MoveNext())
-                    {
-                        list.Add(e.Current);
-                    }
-                }
-
-                return list;
-            }
-
-            public override int GetCount(bool onlyIfCheap)
-            {
-                if (_source is IIListProvider<TSource> listProv)
-                {
-                    int count = listProv.GetCount(onlyIfCheap);
-                    return count == -1 ? -1 : count + _appendCount + _prependCount;
-                }
-
-                return !onlyIfCheap || _source is ICollection<TSource> ? _source.Count() + _appendCount + _prependCount : -1;
             }
         }
     }
