+++ src/System.Linq/src/System/Linq/DefaultIfEmpty.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 
@@ -23,7 +22,7 @@ namespace System.Linq
             return new DefaultIfEmptyIterator<TSource>(source, defaultValue);
         }
 
-        private sealed class DefaultIfEmptyIterator<TSource> : Iterator<TSource>, IIListProvider<TSource>
+        private sealed partial class DefaultIfEmptyIterator<TSource> : Iterator<TSource>
         {
             private readonly IEnumerable<TSource> _source;
             private readonly TSource _default;
@@ -79,38 +78,6 @@ namespace System.Linq
                 }
 
                 base.Dispose();
-            }
-
-            public TSource[] ToArray()
-            {
-                TSource[] array = _source.ToArray();
-                return array.Length == 0 ? new[] { _default } : array;
-            }
-
-            public List<TSource> ToList()
-            {
-                List<TSource> list = _source.ToList();
-                if (list.Count == 0)
-                {
-                    list.Add(_default);
-                }
-
-                return list;
-            }
-
-            public int GetCount(bool onlyIfCheap)
-            {
-                int count;
-                if (!onlyIfCheap || _source is ICollection<TSource> || _source is ICollection)
-                {
-                    count = _source.Count();
-                }
-                else
-                {
-                    count = _source is IIListProvider<TSource> listProv ? listProv.GetCount(onlyIfCheap: true) : -1;
-                }
-
-                return count == 0 ? 1 : count;
             }
         }
     }
