+++ src/System.Linq/src/System/Linq/Enumerable.uap.cs
@@ -8,6 +8,6 @@ namespace System.Linq
 {
     public static partial class Enumerable
     {
-        public static IEnumerable<TSource> AsEnumerable<TSource>(this IEnumerable<TSource> source) => source;
+        public static IEnumerable<TResult> Empty<TResult>() => Array.Empty<TResult>();
     }
 }
