+++ src/System.Linq/src/System/Linq/Range.cs
@@ -19,7 +19,7 @@ namespace System.Linq
 
             if (count == 0)
             {
-                return EmptyPartition<int>.Instance;
+                return Empty<int>();
             }
 
             return new RangeIterator(start, count);
@@ -28,7 +28,7 @@ namespace System.Linq
         /// <summary>
         /// An iterator that yields a range of consecutive integers.
         /// </summary>
-        private sealed class RangeIterator : Iterator<int>, IPartition<int>
+        private sealed partial class RangeIterator : Iterator<int>
         {
             private readonly int _start;
             private readonly int _end;
@@ -67,82 +67,6 @@ namespace System.Linq
             public override void Dispose()
             {
                 _state = -1; // Don't reset current
-            }
-
-            public override IEnumerable<TResult> Select<TResult>(Func<int, TResult> selector)
-            {
-                return new SelectIPartitionIterator<int, TResult>(this, selector);
-            }
-
-            public int[] ToArray()
-            {
-                int[] array = new int[_end - _start];
-                int cur = _start;
-                for (int i = 0; i != array.Length; ++i)
-                {
-                    array[i] = cur;
-                    ++cur;
-                }
-
-                return array;
-            }
-
-            public List<int> ToList()
-            {
-                List<int> list = new List<int>(_end - _start);
-                for (int cur = _start; cur != _end; cur++)
-                {
-                    list.Add(cur);
-                }
-
-                return list;
-            }
-
-            public int GetCount(bool onlyIfCheap) => unchecked(_end - _start);
-
-            public IPartition<int> Skip(int count)
-            {
-                if (count >= _end - _start)
-                {
-                    return EmptyPartition<int>.Instance;
-                }
-
-                return new RangeIterator(_start + count, _end - _start - count);
-            }
-
-            public IPartition<int> Take(int count)
-            {
-                int curCount = _end - _start;
-                if (count >= curCount)
-                {
-                    return this;
-                }
-
-                return new RangeIterator(_start, count);
-            }
-
-            public int TryGetElementAt(int index, out bool found)
-            {
-                if (unchecked((uint)index < (uint)(_end - _start)))
-                {
-                    found = true;
-                    return _start + index;
-                }
-
-                found = false;
-                return 0;
-            }
-
-            public int TryGetFirst(out bool found)
-            {
-                found = true;
-                return _start;
-            }
-
-            public int TryGetLast(out bool found)
-            {
-                found = true;
-                return _end - 1;
             }
         }
     }
