+++ src/System.Linq/src/System/Linq/Range.netcoreapp.cs
@@ -1,0 +1,90 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+
+namespace System.Linq
+{
+    public static partial class Enumerable
+    {
+        private sealed partial class RangeIterator : IPartition<int>
+        {
+            public override IEnumerable<TResult> Select<TResult>(Func<int, TResult> selector)
+            {
+                return new SelectIPartitionIterator<int, TResult>(this, selector);
+            }
+
+            public int[] ToArray()
+            {
+                int[] array = new int[_end - _start];
+                int cur = _start;
+                for (int i = 0; i != array.Length; ++i)
+                {
+                    array[i] = cur;
+                    ++cur;
+                }
+
+                return array;
+            }
+
+            public List<int> ToList()
+            {
+                List<int> list = new List<int>(_end - _start);
+                for (int cur = _start; cur != _end; cur++)
+                {
+                    list.Add(cur);
+                }
+
+                return list;
+            }
+
+            public int GetCount(bool onlyIfCheap) => unchecked(_end - _start);
+
+            public IPartition<int> Skip(int count)
+            {
+                if (count >= _end - _start)
+                {
+                    return EmptyPartition<int>.Instance;
+                }
+
+                return new RangeIterator(_start + count, _end - _start - count);
+            }
+
+            public IPartition<int> Take(int count)
+            {
+                int curCount = _end - _start;
+                if (count >= curCount)
+                {
+                    return this;
+                }
+
+                return new RangeIterator(_start, count);
+            }
+
+            public int TryGetElementAt(int index, out bool found)
+            {
+                if (unchecked((uint)index < (uint)(_end - _start)))
+                {
+                    found = true;
+                    return _start + index;
+                }
+
+                found = false;
+                return 0;
+            }
+
+            public int TryGetFirst(out bool found)
+            {
+                found = true;
+                return _start;
+            }
+
+            public int TryGetLast(out bool found)
+            {
+                found = true;
+                return _end - 1;
+            }
+        }
+    }
+}
