+++ src/System.Linq/src/System/Linq/Repeat.cs
@@ -18,7 +18,7 @@ namespace System.Linq
 
             if (count == 0)
             {
-                return EmptyPartition<TResult>.Instance;
+                return Empty<TResult>();
             }
 
             return new RepeatIterator<TResult>(element, count);
@@ -28,7 +28,7 @@ namespace System.Linq
         /// An iterator that yields the same item multiple times. 
         /// </summary>
         /// <typeparam name="TResult">The type of the item.</typeparam>
-        private sealed class RepeatIterator<TResult> : Iterator<TResult>, IPartition<TResult>
+        private sealed partial class RepeatIterator<TResult> : Iterator<TResult>
         {
             private readonly int _count;
 
@@ -67,77 +67,6 @@ namespace System.Linq
 
                 Dispose();
                 return false;
-            }
-
-            public override IEnumerable<TResult2> Select<TResult2>(Func<TResult, TResult2> selector) =>
-                new SelectIPartitionIterator<TResult, TResult2>(this, selector);
-
-            public TResult[] ToArray()
-            {
-                TResult[] array = new TResult[_count];
-                if (_current != null)
-                {
-                    Array.Fill(array, _current);
-                }
-
-                return array;
-            }
-
-            public List<TResult> ToList()
-            {
-                List<TResult> list = new List<TResult>(_count);
-                for (int i = 0; i != _count; ++i)
-                {
-                    list.Add(_current);
-                }
-
-                return list;
-            }
-
-            public int GetCount(bool onlyIfCheap) => _count;
-
-            public IPartition<TResult> Skip(int count)
-            {
-                if (count >= _count)
-                {
-                    return EmptyPartition<TResult>.Instance;
-                }
-
-                return new RepeatIterator<TResult>(_current, _count - count);
-            }
-
-            public IPartition<TResult> Take(int count)
-            {
-                if (count >= _count)
-                {
-                    return this;
-                }
-
-                return new RepeatIterator<TResult>(_current, count);
-            }
-
-            public TResult TryGetElementAt(int index, out bool found)
-            {
-                if ((uint)index < (uint)_count)
-                {
-                    found = true;
-                    return _current;
-                }
-
-                found = false;
-                return default(TResult);
-            }
-
-            public TResult TryGetFirst(out bool found)
-            {
-                found = true;
-                return _current;
-            }
-
-            public TResult TryGetLast(out bool found)
-            {
-                found = true;
-                return _current;
             }
         }
     }
