+++ src/System.Linq/src/System/Linq/Skip.uap.cs
@@ -1,0 +1,23 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+
+namespace System.Linq
+{
+    public static partial class Enumerable
+    {
+        private static IEnumerable<TSource> SkipIterator<TSource>(IEnumerable<TSource> source, int count)
+        {
+            using (IEnumerator<TSource> e = source.GetEnumerator())
+            {
+                while (count > 0 && e.MoveNext()) count--;
+                if (count <= 0)
+                {
+                    while (e.MoveNext()) yield return e.Current;
+                }
+            }
+        }
+    }
+}
