+++ src/System.Linq/src/System/Linq/Take.cs
@@ -16,22 +16,9 @@ namespace System.Linq
                 throw Error.ArgumentNull(nameof(source));
             }
 
-            if (count <= 0)
-            {
-                return EmptyPartition<TSource>.Instance;
-            }
-
-            if (source is IPartition<TSource> partition)
-            {
-                return partition.Take(count);
-            }
-
-            if (source is IList<TSource> sourceList)
-            {
-                return new ListPartition<TSource>(sourceList, 0, count - 1);
-            }
-
-            return new EnumerablePartition<TSource>(source, 0, count - 1);
+            return count <= 0 ?
+                Empty<TSource>() :
+                TakeIterator<TSource>(source, count);
         }
 
         public static IEnumerable<TSource> TakeWhile<TSource>(this IEnumerable<TSource> source, Func<TSource, bool> predicate)
@@ -103,12 +90,9 @@ namespace System.Linq
                 throw Error.ArgumentNull(nameof(source));
             }
 
-            if (count <= 0)
-            {
-                return EmptyPartition<TSource>.Instance;
-            }
-
-            return TakeLastIterator(source, count);
+            return count <= 0 ?
+                Empty<TSource>() :
+                TakeLastIterator(source, count);
         }
 
         private static IEnumerable<TSource> TakeLastIterator<TSource>(IEnumerable<TSource> source, int count)
