+++ src/System.Linq/src/System/Linq/Take.uap.cs
@@ -1,0 +1,23 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+
+namespace System.Linq
+{
+    public static partial class Enumerable
+    {
+        private static IEnumerable<TSource> TakeIterator<TSource>(IEnumerable<TSource> source, int count)
+        {
+            if (count > 0)
+            {
+                foreach (TSource element in source)
+                {
+                    yield return element;
+                    if (--count == 0) break;
+                }
+            }
+        }
+    }
+}
